first impl; failover and requests beyond incrAndGet() are still missing

+++ src/org/jgroups/blocks/atomic/Counter.java
@@ -9,31 +9,32 @@ public interface Counter {
      * Get the current value of the counter
      * @return The current value
      */
-    public int get();
+    public long get();
 
     /**
      * Sets the counter to a new value
      * @param new_value The new value
      */
-    public void set(int new_value);
+    public void set(long new_value);
 
     /**
      * Atomically updates the counter using a CAS operation
+     *
      * @param expect The expected value of the counter
      * @param update The new value of the counter
      * @return True if the counter could be updated, false otherwise
      */
-    public boolean compareAndSet(int expect, int update);
+    public boolean compareAndSet(long expect, long update);
 
     /**
      * Atomically increments the counter and returns the new value
      * @return The new value
      */
-    public int incrementAndGet();
+    public long incrementAndGet();
 
     /**
      * Atomically decrements the counter and returns the new value
      * @return The new value
      */
-    public int decrementAndGet();
+    public long decrementAndGet();
 }
